+++ integration/tests/acceptance/steps/when_steps.go
@@ -20,9 +20,14 @@ func (tc *TestContext) iCallTheHealthEndpoint() error {
 	return tc.WhenICallHealthEndpoint()
 }
 
+func (tc *TestContext) iCallGetProcessorStatus() error {
+	return tc.WhenICallGetProcessorStatus()
+}
+
 func InitializeWhenSteps(ctx *godog.ScenarioContext, tc *TestContext) {
 	ctx.Step(`^I call the GetStreamConfig endpoint$`, tc.iCallTheGetStreamConfigEndpoint)
 	ctx.Step(`^I call the SyncFeatureFlags endpoint$`, tc.iCallTheSyncFeatureFlagsEndpoint)
 	ctx.Step(`^I wait for flags to be synchronized$`, tc.iWaitForFlagsToBeSynchronized)
 	ctx.Step(`^I call the health endpoint$`, tc.iCallTheHealthEndpoint)
+	ctx.Step(`^I call GetProcessorStatus$`, tc.iCallGetProcessorStatus)
 }